refactor: new update notifier
