+++ lib/Cli/Cli.js
@@ -1,113 +1,22 @@
 const path = require('path');
-const Profiler = require('./Profiler');
-const Store = require('./Store');
-const { NavigatorDirectory } = require('./Navigator');
-const Logger = require('./Logger');
-
-/**
- * Command object.
- */
-class Command {
-    /**
-     * Create a Command instance.
-     *
-     * @param {CLI} scope Command scope (the CLI it will be run from).
-     * @param {string} name Command name.
-     */
-    constructor(scope, name) {
-        this.scope = scope;
-        this.name = name;
-        this.options = [];
-        this.help('');
-    }
-
-    /**
-     * Setter for callback.
-     *
-     * @param {function} callback Callback.
-     * @returns {this}
-     */
-    action(callback) {
-        this.callback = callback;
-        return this;
-    }
-
-    /**
-     * Setter for description.
-     *
-     * @param {string} d Description.
-     * @returns {this}
-     */
-    description(d) {
-        this.desc = d;
-        return this;
-    }
-
-    /**
-     * Setter for help text.
-     *
-     * @param {function|string} callback Help text, or callable that will take care of writing help text.
-     * @returns {this}
-     */
-    help(callback) {
-        if (typeof callback === 'string') {
-            // Wrap help text in a callable.
-            let str = callback;
-            callback = () => {
-                this.scope.logger.newline();
-                this.scope.exec('help', { // Show only relevant portion of help.
-                    lite: true,
-                    commands: [this.name],
-                });
-                if (str) {
-                    this.scope.logger.log(str);
-                    this.scope.logger.newline();
-                }
-            };
-        }
-        this.h = callback;
-        return this;
-    }
-
-    /**
-     * Add an option.
-     *
-     * @param {string} name Option name.
-     * @param {string} description Option description.
-     * @param {boolean} required Is this option required?
-     * @returns {this}
-     */
-    option(name, description, required = false) {
-        this.options.push({
-            name,
-            description,
-            required,
-        });
-        return this;
-    }
-
-    /**
-     * Mark a command as deprecated.
-     *
-     * @param {String} version Since version.
-     * @returns {this}
-     */
-    deprecate(version) {
-        this.deprecated = version;
-        return this;
-    }
-}
+const parse = require('./parse');
+const Command = require('./Command');
+const Profiler = require('../Profiler');
+const Store = require('../Store');
+const { NavigatorDirectory } = require('../Navigator');
+const Logger = require('../Logger');
 
 /**
  * CLI object.
  */
 class CLI {
-    constructor(name, version = '1.0.0') {
-        this.name = name;
-        this.v = version;
+    constructor(binName, packageName, packageVersion = '1.0.0') {
+        this.bin = binName;
+        this.packageName = packageName;
+        this.v = packageVersion;
         this.commands = {};
         this.logger = new Logger();
-        this.store = new Store(this.name);
+        this.store = new Store(packageName);
         this.navigator = new NavigatorDirectory(
             path.resolve(__dirname, '..'),
         );
@@ -138,15 +47,17 @@ class CLI {
     /**
      * CLI entrypoint.
      *
+     * @param {Array<string>} argv A list of arguments.
      * @returns {Promise}
      */
-    async start() {
-        let commands = require('./args');
+    async start(argv) {
+        let commands = parse(argv);
         let update = this.checkUpdate(); // Update check.
         this.options = commands.shift();
         if (this.options['v'] || this.options['version']) {
             // Output version number and exit.
             this.logger.log(this.v);
+            return;
         } else if (commands.length) {
             this.profiler = new Profiler(!!this.options.profile);
 
@@ -170,21 +81,11 @@ class CLI {
         }
 
         try {
-            let remoteVersion = await update;
-            if (remoteVersion) {
-                // Current version don't match with latest available one.
-                let msg = 'RNA CLI update available.';
-                let v = `(${this.v} -> ${remoteVersion})`;
-                let sep = '-------------------------------------------------------';
-                this.logger.newline();
-                this.logger.log(sep);
-                this.logger.newline();
-                this.logger.info(msg, v);
-                this.logger.log('$ npm install -g @chialab/rna-cli');
-                this.logger.log('$ yarn global add @chialab/rna-cli');
-                this.logger.newline();
-                this.logger.log(sep);
-                this.logger.newline();
+            let notifier = await update;
+            if (notifier) {
+                process.on('exit', () => {
+                    this.notifyUpdate();
+                });
             }
         } catch (err) {
             //
@@ -229,6 +130,42 @@ class CLI {
         }
     }
 
+    async notifyUpdate() {
+        let version = this.store.get('update.remote');
+        if (version !== this.v) {
+            this.logger.newline();
+            this.logger.log(''.padEnd(50, '-'));
+            this.logger.newline();
+            this.logger.info(`   ${this.bin} update available`, `${this.v} -> ${version}`);
+            this.logger.newline();
+            this.logger.log(`   $ npm install -g ${this.packageName}`);
+            this.logger.log(`   $ yarn global add ${this.packageName}`);
+            this.logger.newline();
+            this.logger.log(''.padEnd(50, '-'));
+            this.logger.newline();
+            this.store.set('update.check', Date.now());
+        }
+    }
+
+    async checkUpdate() {
+        if (process.env.CI) {
+            return;
+        }
+        let day = 24 * 60 * 60 * 1000;
+        let now = Date.now();
+        try {
+            let remoteVersion = await this.getRemoteVersion();
+            this.store.set('update.remote', remoteVersion);
+        } catch (error) {
+            //
+        }
+        let check = this.store.get('update.check') || 0;
+        if (now - check < day) {
+            return;
+        }
+        return true;
+    }
+
     /**
      * Get latest available version.
      *
@@ -236,14 +173,11 @@ class CLI {
      *
      * @returns {Promise<string>}
      */
-    async checkUpdate() {
-        if (process.env.CI) {
-            return;
-        }
+    async getRemoteVersion() {
         const https = require('https');
         return await new Promise((resolve, reject) => {
             https
-                .get('https://registry.npmjs.org/@chialab%2Frna-cli', (response) => {
+                .get(`https://registry.npmjs.org/${this.packageName}`, (response) => {
                     if (response) {
                         let str = '';
                         response.on('data', (chunk) => {
